backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.118.0",
    "instructor[google-genai]>=1.11.3",
    "matplotlib>=3.10.6",
    "pandas>=2.3.3",
    "pandas-stubs>=2.3.2.250926",
    "pydantic-settings>=2.11.0",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.2",
    "scipy-stubs>=1.16.2.0",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "mypy>=1.18.2",
    "ruff>=0.13.3",
]
